Registration descriptor service implementation part 2

+++ src/net45/WampSharp/WAMP2/V2/MetaApi/Rpc/RegistrationDetails.cs
@@ -12,7 +12,7 @@ namespace WampSharp.V2.MetaApi
         public string Invoke { get; set; }
 
         [DataMember(Name = "id")]
-        public int RegistrationId { get; set; }
+        public long RegistrationId { get; set; }
 
         [DataMember(Name = "match")]
         public string Match { get; set; }
